the refactor continues
