+++ src/pkl/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod internal;
-pub mod non_primitive;
 pub mod pkl_mod;
 mod serializer;
 
+pub use internal::type_constants;
 pub use pkl_mod::PklMod;
 pub use serializer::PklSerialize;
